(PC-4942) split algolia logic in three files
