+++ src/libs/algolia/fetchAlgolia/index.ts
@@ -1,0 +1,1 @@
+export { fetchAlgolia } from './fetchAlgolia'
